Custom title and options for Growl
Ability to change a title to a custom string without cementing it to a TYPE::INFO const. Also, we now can override options from a view with a special ['options'] array. This array will override original settings in main.php layout file.

In order to use a custom title, you have to set a 'customTitle' option to true.

Now setFlash function call looks like this:

\Yii::$app->getSession()->setFlash('Custom TItle', [
                'message' => [
                    'Your message'
                ],
                'customTitle' => true,
                'options' => [
                    'style' => 'success',
                ]
            ]);